T5: read config and initialize database
