+++ build.sbt
@@ -2,8 +2,8 @@ import Dependencies.*
 
 ThisBuild / scalaVersion := "2.13.15"
 ThisBuild / version := "0.1.1-SNAPSHOT"
-ThisBuild / organization := "com.example"
-ThisBuild / organizationName := "example"
+ThisBuild / organization := "nl.rasom.workshop"
+ThisBuild / organizationName := "Scala Workshop"
 
 lazy val agenda = (project in file("."))
   .settings(
@@ -11,5 +11,7 @@ lazy val agenda = (project in file("."))
     libraryDependencies += Zio,
     libraryDependencies += ZioCli,
     libraryDependencies += Sqlite,
+    libraryDependencies += TypeSafeConfig,
+    libraryDependencies += OsLib,
     libraryDependencies += Munit % Test
   )
