+++ src/main/resources/application.conf
@@ -1,0 +1,3 @@
+sqlite { 
+  dbFileRelativePath = ".agenda/agenda.db"
+}
